Restructured the repository for easier usage
