+++ examples/ppo_lstm.py
@@ -13,9 +13,9 @@ import torch.optim as optim
 
 from torch.utils.tensorboard import SummaryWriter
 
-from pytag.gym.wrappers import MergeActionMaskWrapper, RecordEpisodeStatistics
+from utils.wrappers import MergeActionMaskWrapper, RecordEpisodeStatistics
 from pytag.utils.common import make_env
-from utils.networks import PPONet, PPOLSTM
+from utils.networks import PPOLSTM
 
 def parse_args():
     # fmt: off
