+++ examples/utils/networks.py
