+++ pytag/setup.py
@@ -6,5 +6,8 @@ setup(
     packages=find_packages(),
     package_data={'': ['jars/ModernBoardGame.jar']},
     include_package_data=True,
-    install_requires=["gymnasium", "Jpype1", "numpy"]
+    install_requires=["gymnasium", "Jpype1", "numpy"],
+    extras_require={
+        "examples": ["torch", "tensorboard", "wandb"]
+    }
 )
